+++ test/DotSwashbuckle.AspNetCore.SwaggerGen.Test/Fixtures/XmlRemarkAnnotatedTypes.cs
@@ -1,0 +1,26 @@
+﻿namespace DotSwashbuckle.AspNetCore.SwaggerGen.Test.Fixtures
+{
+    /// <summary>
+    /// Summary for XmlRemarkAnnotatedType
+    /// </summary>
+    /// <remarks>
+    /// Remarks for XmlRemarkAnnotatedType
+    /// </remarks>
+    public class XmlAnnotatedSummaryRemarks
+    {
+    }
+
+    /// <summary>
+    /// Summary for XmlRemarkAnnotatedType
+    /// </summary>
+    public class XmlAnnotatedSummary
+    {
+    }
+
+    /// <remarks>
+    /// Remarks for XmlRemarkAnnotatedType
+    /// </remarks>
+    public class XmlAnnotatedRemarks
+    {
+    }
+}